refactor - pending freelancerList Component Split
보류한 프리랜서 리스트와 카드 파일 분리
