+++ src/components/common/MenuTabBarComp.tsx
@@ -42,7 +42,7 @@ const MenuTabBarComp = ({ menu, children }: MenuTabBarCompProps) => {
         {currentTab === "프로젝트 탐색" && <ProjectNavigation />}
         {currentTab === "지원한 프로젝트" && <AppliedProjectList />}
         {currentTab === "제안받은 프로젝트" && <SuggestedProjectList />}
-        {currentTab === "지원한 프리랜서 확인" && <ApplicantFreelancerList />}
+        {currentTab === "지원한 프리랜서" && <ApplicantFreelancerList />}
         {currentTab === "보류한 프리랜서" && <PendingFreelancerList />}
         {currentTab === "우리 기업 구성원" && <MemberList />}
         {currentTab === "진행 중인 프리랜서" && <OngoingFreelancerList />}
